Proposal for using fragments with graphql queries
