+++ client/components/TestRun/queries.js
@@ -1,172 +1,14 @@
 import { gql } from '@apollo/client';
+import {
+    TEST_PLAN_RUN_FIELDS,
+    TEST_PLAN_REPORT_FIELDS
+} from '../common/queries';
 
 export const TEST_RUN_PAGE_QUERY = gql`
+    ${TEST_PLAN_RUN_FIELDS}
     query TestPlanRunPage($testPlanRunId: ID!) {
         testPlanRun(id: $testPlanRunId) {
-            id
-            tester {
-                id
-                username
-            }
-            testResults {
-                id
-                startedAt
-                completedAt
-                test {
-                    id
-                    rowNumber
-                    title
-                    renderedUrl
-                    renderableContent
-                }
-                scenarioResults {
-                    id
-                    scenario {
-                        commands {
-                            id
-                            text
-                        }
-                    }
-                    output
-                    assertionResults {
-                        id
-                        assertion {
-                            text
-                        }
-                        passed
-                    }
-                    requiredAssertionResults: assertionResults(
-                        priority: REQUIRED
-                    ) {
-                        assertion {
-                            text
-                        }
-                        passed
-                    }
-                    optionalAssertionResults: assertionResults(
-                        priority: OPTIONAL
-                    ) {
-                        assertion {
-                            text
-                        }
-                        passed
-                    }
-                    mayAssertionResults: assertionResults(priority: MAY) {
-                        assertion {
-                            text
-                        }
-                        passed
-                    }
-                    unexpectedBehaviors {
-                        id
-                        text
-                        otherUnexpectedBehaviorText
-                    }
-                }
-                atVersion {
-                    id
-                    name
-                }
-                browserVersion {
-                    id
-                    name
-                }
-            }
-            testPlanReport {
-                id
-                conflicts {
-                    source {
-                        test {
-                            id
-                            title
-                            rowNumber
-                        }
-                        scenario {
-                            id
-                            commands {
-                                text
-                            }
-                        }
-                        assertion {
-                            id
-                            text
-                        }
-                    }
-                    conflictingResults {
-                        testPlanRun {
-                            id
-                            tester {
-                                username
-                            }
-                        }
-                        scenarioResult {
-                            output
-                            unexpectedBehaviors {
-                                text
-                                otherUnexpectedBehaviorText
-                            }
-                        }
-                        assertionResult {
-                            passed
-                        }
-                    }
-                }
-                at {
-                    id
-                    name
-                    atVersions {
-                        id
-                        name
-                    }
-                }
-                browser {
-                    id
-                    name
-                    browserVersions {
-                        id
-                        name
-                    }
-                }
-                testPlanVersion {
-                    id
-                    title
-                    phase
-                    updatedAt
-                    gitSha
-                    testPageUrl
-                    testPlan {
-                        directory
-                    }
-                    metadata
-                }
-                runnableTests {
-                    id
-                    rowNumber
-                    title
-                    ats {
-                        id
-                        name
-                    }
-                    atMode
-                    renderedUrl
-                    scenarios {
-                        id
-                        at {
-                            id
-                            name
-                        }
-                        commands {
-                            id
-                            text
-                        }
-                    }
-                    assertions {
-                        id
-                        priority
-                        text
-                    }
-                }
-            }
+            ...TestPlanRunFields
         }
         me {
             id
@@ -181,107 +23,17 @@ export const TEST_RUN_PAGE_QUERY = gql`
 `;
 
 export const TEST_RUN_PAGE_ANON_QUERY = gql`
+    ${TEST_PLAN_REPORT_FIELDS}
     query TestPlanRunAnonPage($testPlanReportId: ID!) {
         testPlanReport(id: $testPlanReportId) {
-            id
-            conflicts {
-                source {
-                    test {
-                        id
-                        title
-                        rowNumber
-                    }
-                    scenario {
-                        id
-                        commands {
-                            text
-                        }
-                    }
-                    assertion {
-                        id
-                        text
-                    }
-                }
-                conflictingResults {
-                    testPlanRun {
-                        id
-                        tester {
-                            username
-                        }
-                    }
-                    scenarioResult {
-                        output
-                        unexpectedBehaviors {
-                            text
-                            otherUnexpectedBehaviorText
-                        }
-                    }
-                    assertionResult {
-                        passed
-                    }
-                }
-            }
-            at {
-                id
-                name
-                atVersions {
-                    id
-                    name
-                }
-            }
-            browser {
-                id
-                name
-                browserVersions {
-                    id
-                    name
-                }
-            }
-            testPlanVersion {
-                id
-                title
-                phase
-                updatedAt
-                gitSha
-                testPageUrl
-                testPlan {
-                    directory
-                }
-                metadata
-            }
-            runnableTests {
-                id
-                rowNumber
-                title
-                ats {
-                    id
-                    name
-                }
-                atMode
-                renderedUrl
-                renderableContent
-                scenarios {
-                    id
-                    at {
-                        id
-                        name
-                    }
-                    commands {
-                        id
-                        text
-                    }
-                }
-                assertions {
-                    id
-                    priority
-                    text
-                }
-            }
+            ...TestPlanReportFields
         }
     }
 `;
 
 export const FIND_OR_CREATE_TEST_RESULT_MUTATION = gql`
+    ${TEST_PLAN_RUN_FIELDS}
+    ${TEST_PLAN_REPORT_FIELDS}
     mutation FindOrCreateTestResult(
         $testPlanRunId: ID!
         $testId: ID!
@@ -296,266 +48,10 @@ export const FIND_OR_CREATE_TEST_RESULT_MUTATION = gql`
             ) {
                 locationOfData
                 testPlanRun {
-                    id
-                    tester {
-                        id
-                        username
-                    }
-                    testResults {
-                        id
-                        startedAt
-                        completedAt
-                        test {
-                            id
-                            rowNumber
-                            title
-                            renderedUrl
-                            renderableContent
-                        }
-                        scenarioResults {
-                            id
-                            scenario {
-                                commands {
-                                    id
-                                    text
-                                }
-                            }
-                            output
-                            assertionResults {
-                                id
-                                assertion {
-                                    text
-                                }
-                                passed
-                            }
-                            requiredAssertionResults: assertionResults(
-                                priority: REQUIRED
-                            ) {
-                                assertion {
-                                    text
-                                }
-                                passed
-                            }
-                            optionalAssertionResults: assertionResults(
-                                priority: OPTIONAL
-                            ) {
-                                assertion {
-                                    text
-                                }
-                                passed
-                            }
-                            mayAssertionResults: assertionResults(
-                                priority: MAY
-                            ) {
-                                assertion {
-                                    text
-                                }
-                                passed
-                            }
-                            unexpectedBehaviors {
-                                id
-                                text
-                                otherUnexpectedBehaviorText
-                            }
-                        }
-                        atVersion {
-                            id
-                            name
-                        }
-                        browserVersion {
-                            id
-                            name
-                        }
-                    }
-                    testPlanReport {
-                        id
-                        conflicts {
-                            source {
-                                test {
-                                    id
-                                    title
-                                    rowNumber
-                                }
-                                scenario {
-                                    id
-                                    commands {
-                                        text
-                                    }
-                                }
-                                assertion {
-                                    id
-                                    text
-                                }
-                            }
-                            conflictingResults {
-                                testPlanRun {
-                                    id
-                                    tester {
-                                        username
-                                    }
-                                }
-                                scenarioResult {
-                                    output
-                                    unexpectedBehaviors {
-                                        text
-                                        otherUnexpectedBehaviorText
-                                    }
-                                }
-                                assertionResult {
-                                    passed
-                                }
-                            }
-                        }
-                        at {
-                            id
-                            name
-                            atVersions {
-                                id
-                                name
-                            }
-                        }
-                        browser {
-                            id
-                            name
-                            browserVersions {
-                                id
-                                name
-                            }
-                        }
-                        testPlanVersion {
-                            id
-                            title
-                            phase
-                            updatedAt
-                            gitSha
-                            testPageUrl
-                            testPlan {
-                                directory
-                            }
-                            metadata
-                        }
-                        runnableTests {
-                            id
-                            rowNumber
-                            title
-                            ats {
-                                id
-                                name
-                            }
-                            atMode
-                            renderedUrl
-                            scenarios {
-                                id
-                                at {
-                                    id
-                                    name
-                                }
-                                commands {
-                                    id
-                                    text
-                                }
-                            }
-                            assertions {
-                                id
-                                priority
-                                text
-                            }
-                        }
-                    }
+                    ...TestPlanRunFields
                 }
                 testPlanReport {
-                    id
-                    conflicts {
-                        source {
-                            test {
-                                id
-                                title
-                                rowNumber
-                            }
-                            scenario {
-                                id
-                                commands {
-                                    text
-                                }
-                            }
-                            assertion {
-                                id
-                                text
-                            }
-                        }
-                        conflictingResults {
-                            testPlanRun {
-                                id
-                                tester {
-                                    username
-                                }
-                            }
-                            scenarioResult {
-                                output
-                                unexpectedBehaviors {
-                                    text
-                                    otherUnexpectedBehaviorText
-                                }
-                            }
-                            assertionResult {
-                                passed
-                            }
-                        }
-                    }
-                    at {
-                        id
-                        name
-                        atVersions {
-                            id
-                            name
-                        }
-                    }
-                    browser {
-                        id
-                        name
-                        browserVersions {
-                            id
-                            name
-                        }
-                    }
-                    testPlanVersion {
-                        id
-                        title
-                        phase
-                        gitSha
-                        testPageUrl
-                        testPlan {
-                            directory
-                        }
-                        metadata
-                    }
-                    runnableTests {
-                        id
-                        rowNumber
-                        title
-                        ats {
-                            id
-                            name
-                        }
-                        atMode
-                        renderedUrl
-                        scenarios {
-                            id
-                            at {
-                                id
-                                name
-                            }
-                            commands {
-                                id
-                                text
-                            }
-                        }
-                        assertions {
-                            id
-                            priority
-                            text
-                        }
-                    }
+                    ...TestPlanReportFields
                 }
             }
         }
@@ -563,6 +59,8 @@ export const FIND_OR_CREATE_TEST_RESULT_MUTATION = gql`
 `;
 
 export const SAVE_TEST_RESULT_MUTATION = gql`
+    ${TEST_PLAN_RUN_FIELDS}
+    ${TEST_PLAN_REPORT_FIELDS}
     mutation SaveTestResult(
         $id: ID!
         $atVersionId: ID!
@@ -580,266 +78,10 @@ export const SAVE_TEST_RESULT_MUTATION = gql`
             ) {
                 locationOfData
                 testPlanRun {
-                    id
-                    tester {
-                        id
-                        username
-                    }
-                    testResults {
-                        id
-                        startedAt
-                        completedAt
-                        test {
-                            id
-                            rowNumber
-                            title
-                            renderedUrl
-                            renderableContent
-                        }
-                        scenarioResults {
-                            id
-                            scenario {
-                                commands {
-                                    id
-                                    text
-                                }
-                            }
-                            output
-                            assertionResults {
-                                id
-                                assertion {
-                                    text
-                                }
-                                passed
-                            }
-                            requiredAssertionResults: assertionResults(
-                                priority: REQUIRED
-                            ) {
-                                assertion {
-                                    text
-                                }
-                                passed
-                            }
-                            optionalAssertionResults: assertionResults(
-                                priority: OPTIONAL
-                            ) {
-                                assertion {
-                                    text
-                                }
-                                passed
-                            }
-                            mayAssertionResults: assertionResults(
-                                priority: MAY
-                            ) {
-                                assertion {
-                                    text
-                                }
-                                passed
-                            }
-                            unexpectedBehaviors {
-                                id
-                                text
-                                otherUnexpectedBehaviorText
-                            }
-                        }
-                        atVersion {
-                            id
-                            name
-                        }
-                        browserVersion {
-                            id
-                            name
-                        }
-                    }
-                    testPlanReport {
-                        id
-                        conflicts {
-                            source {
-                                test {
-                                    id
-                                    title
-                                    rowNumber
-                                }
-                                scenario {
-                                    id
-                                    commands {
-                                        text
-                                    }
-                                }
-                                assertion {
-                                    id
-                                    text
-                                }
-                            }
-                            conflictingResults {
-                                testPlanRun {
-                                    id
-                                    tester {
-                                        username
-                                    }
-                                }
-                                scenarioResult {
-                                    output
-                                    unexpectedBehaviors {
-                                        text
-                                        otherUnexpectedBehaviorText
-                                    }
-                                }
-                                assertionResult {
-                                    passed
-                                }
-                            }
-                        }
-                        at {
-                            id
-                            name
-                            atVersions {
-                                id
-                                name
-                            }
-                        }
-                        browser {
-                            id
-                            name
-                            browserVersions {
-                                id
-                                name
-                            }
-                        }
-                        testPlanVersion {
-                            id
-                            title
-                            phase
-                            updatedAt
-                            gitSha
-                            testPageUrl
-                            testPlan {
-                                directory
-                            }
-                            metadata
-                        }
-                        runnableTests {
-                            id
-                            rowNumber
-                            title
-                            ats {
-                                id
-                                name
-                            }
-                            atMode
-                            renderedUrl
-                            scenarios {
-                                id
-                                at {
-                                    id
-                                    name
-                                }
-                                commands {
-                                    id
-                                    text
-                                }
-                            }
-                            assertions {
-                                id
-                                priority
-                                text
-                            }
-                        }
-                    }
+                    ...TestPlanRunFields
                 }
                 testPlanReport {
-                    id
-                    conflicts {
-                        source {
-                            test {
-                                id
-                                title
-                                rowNumber
-                            }
-                            scenario {
-                                id
-                                commands {
-                                    text
-                                }
-                            }
-                            assertion {
-                                id
-                                text
-                            }
-                        }
-                        conflictingResults {
-                            testPlanRun {
-                                id
-                                tester {
-                                    username
-                                }
-                            }
-                            scenarioResult {
-                                output
-                                unexpectedBehaviors {
-                                    text
-                                    otherUnexpectedBehaviorText
-                                }
-                            }
-                            assertionResult {
-                                passed
-                            }
-                        }
-                    }
-                    at {
-                        id
-                        name
-                        atVersions {
-                            id
-                            name
-                        }
-                    }
-                    browser {
-                        id
-                        name
-                        browserVersions {
-                            id
-                            name
-                        }
-                    }
-                    testPlanVersion {
-                        id
-                        title
-                        phase
-                        gitSha
-                        testPageUrl
-                        testPlan {
-                            directory
-                        }
-                        metadata
-                    }
-                    runnableTests {
-                        id
-                        rowNumber
-                        title
-                        ats {
-                            id
-                            name
-                        }
-                        atMode
-                        renderedUrl
-                        scenarios {
-                            id
-                            at {
-                                id
-                                name
-                            }
-                            commands {
-                                id
-                                text
-                            }
-                        }
-                        assertions {
-                            id
-                            priority
-                            text
-                        }
-                    }
+                    ...TestPlanReportFields
                 }
             }
         }
@@ -847,6 +89,8 @@ export const SAVE_TEST_RESULT_MUTATION = gql`
 `;
 
 export const SUBMIT_TEST_RESULT_MUTATION = gql`
+    ${TEST_PLAN_RUN_FIELDS}
+    ${TEST_PLAN_REPORT_FIELDS}
     mutation SubmitTestResult(
         $id: ID!
         $atVersionId: ID!
@@ -864,266 +108,10 @@ export const SUBMIT_TEST_RESULT_MUTATION = gql`
             ) {
                 locationOfData
                 testPlanRun {
-                    id
-                    tester {
-                        id
-                        username
-                    }
-                    testResults {
-                        id
-                        startedAt
-                        completedAt
-                        test {
-                            id
-                            rowNumber
-                            title
-                            renderedUrl
-                            renderableContent
-                        }
-                        scenarioResults {
-                            id
-                            scenario {
-                                commands {
-                                    id
-                                    text
-                                }
-                            }
-                            output
-                            assertionResults {
-                                id
-                                assertion {
-                                    text
-                                }
-                                passed
-                            }
-                            requiredAssertionResults: assertionResults(
-                                priority: REQUIRED
-                            ) {
-                                assertion {
-                                    text
-                                }
-                                passed
-                            }
-                            optionalAssertionResults: assertionResults(
-                                priority: OPTIONAL
-                            ) {
-                                assertion {
-                                    text
-                                }
-                                passed
-                            }
-                            mayAssertionResults: assertionResults(
-                                priority: MAY
-                            ) {
-                                assertion {
-                                    text
-                                }
-                                passed
-                            }
-                            unexpectedBehaviors {
-                                id
-                                text
-                                otherUnexpectedBehaviorText
-                            }
-                        }
-                        atVersion {
-                            id
-                            name
-                        }
-                        browserVersion {
-                            id
-                            name
-                        }
-                    }
-                    testPlanReport {
-                        id
-                        conflicts {
-                            source {
-                                test {
-                                    id
-                                    title
-                                    rowNumber
-                                }
-                                scenario {
-                                    id
-                                    commands {
-                                        text
-                                    }
-                                }
-                                assertion {
-                                    id
-                                    text
-                                }
-                            }
-                            conflictingResults {
-                                testPlanRun {
-                                    id
-                                    tester {
-                                        username
-                                    }
-                                }
-                                scenarioResult {
-                                    output
-                                    unexpectedBehaviors {
-                                        text
-                                        otherUnexpectedBehaviorText
-                                    }
-                                }
-                                assertionResult {
-                                    passed
-                                }
-                            }
-                        }
-                        at {
-                            id
-                            name
-                            atVersions {
-                                id
-                                name
-                            }
-                        }
-                        browser {
-                            id
-                            name
-                            browserVersions {
-                                id
-                                name
-                            }
-                        }
-                        testPlanVersion {
-                            id
-                            title
-                            phase
-                            updatedAt
-                            gitSha
-                            testPageUrl
-                            testPlan {
-                                directory
-                            }
-                            metadata
-                        }
-                        runnableTests {
-                            id
-                            rowNumber
-                            title
-                            ats {
-                                id
-                                name
-                            }
-                            atMode
-                            renderedUrl
-                            scenarios {
-                                id
-                                at {
-                                    id
-                                    name
-                                }
-                                commands {
-                                    id
-                                    text
-                                }
-                            }
-                            assertions {
-                                id
-                                priority
-                                text
-                            }
-                        }
-                    }
+                    ...TestPlanRunFields
                 }
                 testPlanReport {
-                    id
-                    conflicts {
-                        source {
-                            test {
-                                id
-                                title
-                                rowNumber
-                            }
-                            scenario {
-                                id
-                                commands {
-                                    text
-                                }
-                            }
-                            assertion {
-                                id
-                                text
-                            }
-                        }
-                        conflictingResults {
-                            testPlanRun {
-                                id
-                                tester {
-                                    username
-                                }
-                            }
-                            scenarioResult {
-                                output
-                                unexpectedBehaviors {
-                                    text
-                                    otherUnexpectedBehaviorText
-                                }
-                            }
-                            assertionResult {
-                                passed
-                            }
-                        }
-                    }
-                    at {
-                        id
-                        name
-                        atVersions {
-                            id
-                            name
-                        }
-                    }
-                    browser {
-                        id
-                        name
-                        browserVersions {
-                            id
-                            name
-                        }
-                    }
-                    testPlanVersion {
-                        id
-                        title
-                        phase
-                        gitSha
-                        testPageUrl
-                        testPlan {
-                            directory
-                        }
-                        metadata
-                    }
-                    runnableTests {
-                        id
-                        rowNumber
-                        title
-                        ats {
-                            id
-                            name
-                        }
-                        atMode
-                        renderedUrl
-                        scenarios {
-                            id
-                            at {
-                                id
-                                name
-                            }
-                            commands {
-                                id
-                                text
-                            }
-                        }
-                        assertions {
-                            id
-                            priority
-                            text
-                        }
-                    }
+                    ...TestPlanReportFields
                 }
             }
         }
